move /global to /api/global
implement #337
