+++ example/namedtracer/foo/foo.go
@@ -17,9 +17,9 @@ package foo
 import (
 	"context"
 
+	"go.opentelemetry.io/otel/api/global"
 	"go.opentelemetry.io/otel/api/key"
 	"go.opentelemetry.io/otel/api/trace"
-	"go.opentelemetry.io/otel/global"
 )
 
 var (
